Replace `ws-butler` with `whitespace-cleanup-mode`
Since the former seems to modify text properties, resulting in some
weird shit going on like in the debugger I can see that even the first
M-x command is passed wrapped like `#("command" 0 40 (ws-butler-chg
chg))` and I can't fix it. It broke `straight-freeze-versions` because
it would write that text into the buffer.