Fix highlight face
Finally tracked this down! Since I open Emacs as a server, I was
accidentally loading the ‘tango-dark’ theme, and then when the window
opened, my Solarized theme was applied on top. That almost worked
perfectly, except that ‘tango-dark’ customized the foreground face of
‘highlight’, hence my weird issue that only showed up here.

Also applied an ordering to ‘solarized-andy’ so that the ‘themes’ folder
is created in the right place.