Remove `M-[` binding which breaks terminal Emacs
Emacs 25 and up automatically enables “bracketed paste mode” which is
great because Emacs can tell when something is pasted into it. However,
by having a binding which conflicts with something in
`input-decode-map`, such as escape on `M-[`, this support breaks and the
characters used to support bracketed paste are misinterpreted. Thus the
pasted text is surrounded with “200~” etc. It is strange, but removing
the problematic binding fixes it.